Extract utils module
